add progress bar and break up logic, process es
