+++ process_dictionaries/utils/processWordFrequencies.py
@@ -1,9 +1,3 @@
-import wordfreq as wf
-from PyMultiDictionary import MultiDictionary, DICT_EDUCALINGO
-import urllib.error
-import json
-import os
-
 '''
 Given the file with the dictionary created by processWordDump(), this component will use the wordfreq module to make a more suitable word database for the Talking Circles game.
 Method:
@@ -19,6 +13,14 @@ Method:
         - If keyboard interrupt, move on to next top word without adding to dictionary
 3. Write list of words within frequency range to new file
 '''
+
+import urllib.error
+import json
+import os
+
+import wordfreq as wf
+from alive_progress import alive_bar
+from PyMultiDictionary import MultiDictionary, DICT_EDUCALINGO
 
 def processWordFrequencies(draftFile, roughFile, languageCode):
   dictionary = MultiDictionary()
@@ -40,47 +42,45 @@ def processWordFrequencies(draftFile, roughFile, languageCode):
   with open(draftFile, "r", encoding="utf-8") as f:
     for lexicon in f: 
       unfilteredWordDict = json.loads(lexicon)
-      wordCount = 0
 
-      for commonWord in commonWords:
+      with alive_bar(10000, title="Sorting by frequency", stats=False) as bar:
+        for commonWord in commonWords:
 
-        wordCount += 1
-        if (wordCount % 100 == 0):
-          print(f'{wordCount // 100}% processed')
+          for unfilteredWord in unfilteredWordDict:
+            if commonWord == unfilteredWord:
+              if unfilteredWord in words:
+                break # Go to next common word, this one is already in the out file
 
-        for unfilteredWord in unfilteredWordDict:
-          if commonWord == unfilteredWord:
-            if unfilteredWord in words:
-              break # Go to next common word, this one is already in the out file
+              while True:
+                try:
+                  words[commonWord] = unfilteredWordDict[unfilteredWord]
+                  words[commonWord]["frequency"] = frequencyDict[commonWord]
 
-            while True:
-              try:
-                words[commonWord] = unfilteredWordDict[unfilteredWord]
-                words[commonWord]["frequency"] = frequencyDict[commonWord]
+                  if eduWords:
+                    EduSynonyms = dictionary.synonym(languageCode, commonWord, dictionary=DICT_EDUCALINGO)
 
-                if eduWords:
-                  EduSynonyms = dictionary.synonym(languageCode, commonWord, dictionary=DICT_EDUCALINGO)
-
-                  if len(EduSynonyms) > 0:
-                    for synonym in EduSynonyms:
-                      if synonym.casefold() not in words[commonWord]["word"] and \
-                        words[commonWord]["word"] not in synonym.casefold() and \
-                        not any(synonym.casefold() in similarWord[1].casefold() for similarWord in words[commonWord]["related words"]):
-                        words[commonWord]["related words"].append(["synonym", synonym])
-                      
-                break # Out of the while loop
-
-              except urllib.error.URLError:
-                print(f'Failed to get synonyms for {commonWord}')
-                ans = input("(y/n) Keep trying? ")
-                if ans == 'n':
+                    if len(EduSynonyms) > 0:
+                      for synonym in EduSynonyms:
+                        if synonym.casefold() not in words[commonWord]["word"] and \
+                          words[commonWord]["word"] not in synonym.casefold() and \
+                          not any(synonym.casefold() in similarWord[1].casefold() for similarWord in words[commonWord]["related words"]):
+                          words[commonWord]["related words"].append(["synonym", synonym])
+                        
                   break # Out of the while loop
 
-              except KeyboardInterrupt:
-                print(f'Got stuck getting synonyms for {commonWord}')
-                break
+                except urllib.error.URLError:
+                  print(f'Failed to get synonyms for {commonWord}')
+                  ans = input("(y/n) Keep trying? ")
+                  if ans == 'n':
+                    break # Out of the while loop
 
-            break # Go to next common word, this one is found
+                except KeyboardInterrupt:
+                  print(f'Got stuck getting synonyms for {commonWord}')
+                  break
+
+              break # Go to next common word, this one is found
+
+          bar()
 
   r_file = open(roughFile, 'w+', encoding="utf-8")
   json.dump(words, r_file)
